Create aesFinalidea
This is how I will be implementing my AES into the database for each user. Each user will have their own key. And for every new encryption, a new IV will be generated. And when decrypting, the user has to use their own specific key and that specific IV to decrypt the message. The IV will be kept in the same database as the ciphertext passwords.